requirements.txt
Flask==3.0.0
Pillow==10.1.0
openslide-python==1.3.1